+++ bon.immutables/src/main/java/com/microej/example/foundation/bon/immutables/ExampleImmutables.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2009-2023 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.foundation.bon.immutables;