+++ drawing.antialiased/src/main/java/com/microej/example/foundation/microui/antialiased/ExampleAntialiased.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2020 MicroEJ Corp. All rights reserved.
+ * Copyright 2014-2023 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.foundation.microui.antialiased;