+++ CHANGELOG.rst
@@ -8,6 +8,16 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 -------------------
+51.0.0 - 2024-01-29
+-------------------
+
+Changed
+=======
+
+- Update to MICROEJ SDK 6 and switch from Ivy to Gradle.
+- Update READMEs and LICENSEs.
+
+-------------------
 50.0.0 - 2023-07-13
 -------------------
 
@@ -160,5 +170,5 @@ MicroEj 3.1:
 --------------
 
 .. ReStructuredText
-.. Copyright 2020-2023 MicroEJ Corp. All rights reserved.
+.. Copyright 2020-2024 MicroEJ Corp. All rights reserved.
 .. Use of this source code is governed by a BSD-style license that can be found with this software.