+++ drawing.antialiased/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.0.0 - 2024-01-29
+
+### Changed
+
+- Update to MICROEJ SDK 6 and switch from Ivy to Gradle.
+- Update libraries version.
+- Update README and LICENSE.
+
 ## 3.0.0 - 2020-12-14
 
 ### Added
@@ -49,5 +57,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  
 ---  
 _Markdown_   
-_Copyright 2016-2020 MicroEJ Corp. All rights reserved._  
+_Copyright 2016-2024 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._